requirements.txt
streamlit
Pypdf2
google.generativeai
python-dotenv